[core] Move UsageStatsClient to gcs_server/.
