+++ BUILD.bazel
@@ -1888,17 +1888,33 @@ ray_cc_test(
     ],
 )
 
+
+ray_cc_test(
+    name = "usage_stats_client_test",
+    size = "small",
+    srcs = [
+        "src/ray/gcs/gcs_server/test/usage_stats_client_test.cc",
+    ],
+    tags = ["team:core"],
+    deps = [
+        ":gcs_server_lib",
+        ":gcs_server_test_util",
+        ":gcs_test_util_lib",
+        ":ray_mock",
+        "@com_google_googletest//:gtest_main",
+    ],
+)
+
+
 ray_cc_library(
     name = "gcs_client_lib",
     srcs = [
         "src/ray/gcs/gcs_client/accessor.cc",
         "src/ray/gcs/gcs_client/gcs_client.cc",
-        "src/ray/gcs/gcs_client/usage_stats_client.cc",
     ],
     hdrs = [
         "src/ray/gcs/gcs_client/accessor.h",
         "src/ray/gcs/gcs_client/gcs_client.h",
-        "src/ray/gcs/gcs_client/usage_stats_client.h",
     ],
     deps = [
         ":gcs",
@@ -2009,20 +2025,6 @@ ray_cc_test(
         "no_windows",
         "team:core",
     ],
-    deps = [
-        ":gcs_client_lib",
-        ":gcs_server_lib",
-        ":gcs_test_util_lib",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
-    name = "usage_stats_client_test",
-    srcs = [
-        "src/ray/gcs/gcs_client/test/usage_stats_client_test.cc",
-    ],
-    tags = ["team:core"],
     deps = [
         ":gcs_client_lib",
         ":gcs_server_lib",
