+++ src/mock/ray/gcs/gcs_server/gcs_kv_manager.h
@@ -12,6 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include "gmock/gmock.h"
+#include "ray/gcs/gcs_server/gcs_kv_manager.h"
+
 namespace ray {
 namespace gcs {
