+++ src/ray/gcs/gcs_server/gcs_task_manager.h
@@ -18,7 +18,8 @@
 #include "absl/container/flat_hash_map.h"
 #include "absl/container/flat_hash_set.h"
 #include "absl/synchronization/mutex.h"
-#include "ray/gcs/gcs_client/usage_stats_client.h"
+#include "ray/common/asio/periodical_runner.h"
+#include "ray/gcs/gcs_server/usage_stats_client.h"
 #include "ray/gcs/pb_util.h"
 #include "ray/rpc/gcs_server/gcs_rpc_server.h"
 #include "ray/util/counter_map.h"
