+++ src/ray/gcs/gcs_server/gcs_worker_manager.h
@@ -14,10 +14,9 @@
 
 #pragma once
 
-#include "ray/gcs/gcs_client/gcs_client.h"
-#include "ray/gcs/gcs_client/usage_stats_client.h"
 #include "ray/gcs/gcs_server/gcs_kv_manager.h"
 #include "ray/gcs/gcs_server/gcs_table_storage.h"
+#include "ray/gcs/gcs_server/usage_stats_client.h"
 #include "ray/gcs/pubsub/gcs_pub_sub.h"
 #include "ray/rpc/gcs_server/gcs_rpc_server.h"
